Fixed bug in plotting isoforms
A very old bug existed where isoforms of genes would not be plotted.
This has apparently existed for 4-5 years, and was only the product of a
single variable name not being used correctly. We rarely (read: never)
make plots with isoforms, so not too surprising this went unnoticed for
so long.